metrics: system.events is UInt64, scan as uint64
The driver refuses to coerce UInt64 columns into *int64 ("try using
*uint64"), which surfaced when chtop polled system.events on a real
server. Switch the Events map to map[string]uint64 with a dedicated
scanUint64Pairs helper; perSecondRate now takes uint64 too. Tests
adjusted.

+++ internal/tui/metrics_view.go
@@ -176,7 +176,7 @@ func (v *MetricsView) render() {
 		line := fmt.Sprintf(
 			"  %s  %s",
 			muted.Render(padRight(k, 24)),
-			val.Render(padRight(formatMetric(k, curr), 14)),
+			val.Render(padRight(formatMetric(k, int64(curr)), 14)),
 		)
 		if v.prev.SampledAt.IsZero() {
 			line += "  " + muted.Render("(rate next tick)")
